Wait for ajax calls to finish before waiting for css property
The deletion popup takes a bit of time to disappear, after confirming for deletion.
We want to wait for the popup to disappear, and successful deletion of file in the server.

Also, we want to maximize time before the popup disappears. So, reversing the order of
wait provides us maximum waiting.