Add missing mixin for formDateFromNow in public link item
After refactoring the file list item into PublicLinkListItem, the method
formDateFromNow was not accessible any more as it was present in the
mixin from the original component.

This fix adds the mixing to PublicLinkListItem to make that function
work again and fixes the issue where links that have an expiration date
set would not be shown.