Remove store mutations when adding/editing collaborators or link shares
- removed triggers for busy-state for collaborator and link saving/creating.
using local variable in respective form components now. This fixes the
animations in the sidebar panels when adding or removing items, as the
loader is not triggered anymore. As saving can be represented within the
respective form, there is no need to push this state into the store.
- removed unused code in collaborator editing form
- changed buttons in collaborator forms to represent a saving-state when still in
saving mode
- re-introduced a data change detection for collaborators.
- switched promise handling when adding collaborators to using p-queue