Update share indicator in file row after changes
Whenever changes are made in the share panel, the share indicator in the
currently highlighted file list row is now properly updated.

Shares and links data structure used to be separated but they really come from the same API calls.
In this commit we have merged the data structures together into a single state array,
but kept the accessors to deliver filtered results to be able to split between user/group/remote shares and link shares.

Co-authored-by: Benedikt Kulmann <benedikt@kulmann.biz>