Fix passing $gettext to pendingShare action
The mistake was passing in the result of $gettext() instead of the
function itself. This caused `make l10n-read` to fail because of the
missing argument.

For consistency, this commit also adjusts the function name used in the
action to also be called $gettext, to make sure the text to translate
will be found by the l10n tools as well.