Add popup with reshare information + refactor share object
Added popup with reshare information which lists the resharers, their
avatar and additional info.

Refactored the share/collaborator object to have a clearer distinction
between recipient/collaborator, owner and file owner, which eases the
manipulation that is needed in some places where all fields are copied
over for the resharers and owner.