Use alternative temporary ids for uniqueness in views
When using OCIS as backend, the ids of resources is a string instead of
integer. So we cannot embed those into DOM node ids and need to use
another alternative. This fix introduces a unique listId which is only
there to provide uniqueness across the current list and should not be
used for any data related operation.

This fixes the file actions menu when using OCIS as backend.