Do not handle file.sdate in FileDetails
I could not find any context in which this is actually set - neither in
oCIS nor in oC 10. mdate on the other hand was always available.

The shared date is retrieved via sharesTree anyway.
So when we provide sdate and a sharesTree in a unit test
we see the shared date twice, which hardly makes sense.