Fix import of sanitize-html in unit tests (on my machine)
...

I've spent quite some time with @fschadei and @pascalwengerter debugging
this and we are going insane over this.
I still cannot make much sense of this. This change fixes it reliably on
my machine.

If I "import * as sanitizeHtml from 'sanitize-html'", I can access
the function directly in tests or via the .default property.
In rollup builds I can only access it via the .default property.

That means: .default works in both environments, so I would like
to go for this for now. If you have better solutions, I'm very curious
to hear.

Without this change we sometimes (in certain environments that
we could not pin down) saw the following error while running unit or
integration tests:

sanitize_html_1.default is not a function

The weirdest part of this was that I could reliably reproduce this
in my usual development dir, but not anymore when I changed the path
from .../web to .../web2 instead (with fresh checkouts or renaming
the folder, didn't matter).

@pascalwengerter could confirm this for two other paths on his
machine. We saw the error in CI, @fschade could not reproduce it all.

I've seen a lot of weird issues, but this is one of the weirdest
problems I've ever had with computers ... ... ... ... ...