Include current route query in breadcrumb items
As of now we want to keep all query options active when navigating
up/down a folder hierarchy (sorting, pagination options, and now also
the shareId for the SharedResource view). We even have code to restore
query options if they are not set. This commit fixes navigation in
SharedResource breadcrumbs (those NEED the shareId, otherwise we can't
construct the correct webdav path at the moment) and also reduces the
occurrences of reloading query options from the local storage (because
they are already set via the breadcrumb now).