useAppFileHandling.getUrlForResource: add disposition option
which can be either inline or attachment and can be used to define the purpose of the retrieved url.
Chrome allows us to use signed/downloadURLs for video playback although they send a Content-Disposition header, so we can just use a signed url and don't need to create a blob url for this.
We cannot embed PDFs in the same way, as they trigger downloads. So we need to create a blob url for those.

If you need a download url, it's fine (or even desired) for the response
to contain the Content-Disposition header.