perf: simplify table selection-checkbox model
The checkbox for selection in the ResourceTable had the selected
resources as model, which lead to the virtual dom of all checkboxes
being invalidated if the selection got changed at all (adding / removing
one item to/from the selection). Changed the checkbox model to only
reflect the selection state of the one resource it is about. With this
only the virtual dom of the checkboxes gets invlidated that were added
to or removed from the selection.