Shared with me/others/links views
Shares will always appear as a space, which breaks opening single files shares with apps. It breaks because web was supposed to redirect to the universal url (all files), instead of a "share" url. I consider this to be broken in upstream, as this redirection was supposed to be supported.