Auth redirection workaround
The issue was that users were losing the redirection state when they need to login again (go to the SSO and back). A proper fix needs to keep track of original destination in all cases, including the redirection to the SSO from a public link.

To test this issue, we just need to "break" the refresh token and update the expiration of the auth token, forcing web to re-authenticate.