renovatebot: ignore @adobe/leonardo-contrast-colors
The package does not ship a commonjs build anymore and we would need to refactor our build-tokens script to be able to async import it.
It's quite a bit of effort and as the script is just a build time depedency and will probably never see any relevant security fix,
we stick to the old version for now. Otherwise the upgrade to version 1.0.0-alpha.16 will break all renovatebot PRs.