feat: cloud import conflict handling
Adds conflict handling to the recently added cloud import feature.

The import process has been refactored and moved to the HandleUpload Uppy plugin in the course of this. This also introduces a quota check as well as the possibility to have the directory tree created. On top of that it also gathers most of the upload logic in one place instead of having it spread all over the place like before. As a result, the custom drop plugin and the useUploadHelpers composables were removed.