Do not open OcSelect Dropdown on focus
When vue-select gets focus, it immediately opens its dropdown. This is very odd in a formular, where the first item is a oc-select.
Apart from that it behaves differently from native select boxes and usually it's a good idea to stick as closely as possible to native behaviors if native components can't be used for whatever reason.