refactor: resource conflict modal
Refactors the resource conflict modal to use a custom component. This allows us to remove the secondary action as well as the checkbox form the global modal as they were overly specific. The modal itself should be as slick as possible, anything custom like in this case should be handled by a custom component.