refactor: move capabilities to pinia store
Moves the capabilities from vuex store to pinia. Also removes the `use...Capability` composables in favor of the pinia store composable. Furthermore this guarantees that we now have one single location that defines the capabilities and their default values. Before this was happening in the vuex store as well as in the `use...Capability` composable.

Note that this is a breaking change for developers!