fix: replacing folders
Fixes the "Replace" conflict option when copying/moving a folder.

The mechanism of `isOverwritingParentFolder` falsely prevented this though it should only care about the edge case scenario when moving a sub-folder with the same name as the current folder's parent folder into the parent (via breadcrumbs). This needs an additional check if the current folder differs from the target location.