fix: active share nav item
Fixes an issue where the share nav item would be wrongly marked as active after closing a file viewer/editor in the personal space. This happened because the `currentSpace` had been reset as soon as the `useDriveResolver` composable got unmounted. This must not happen in personal or project spaces though.