feat: make unsynced shares accessible
Makes unsynced shares accessible. The way to make this work is by using the remote item id instead of the share jail id for opening/resolving into incoming shares.

In the future we should rethink the `shareId` property since the name is not really fitting anymore. `remoteItemId` would be more fitting.