refactor: rename shareId and shareRoot
Renames `shareId` to `remoteItemId` and `shareRoot` to `remoteItemPath` on resources to match with the Graph terminology.

Also removes the `remoteItemId` on space resources because it only matters for share spaces, in which case this is simply the space id. That means instead of blindly going for `space.remoteItemId`, it now checks if the space is actually a share space and then uses its id.