feat: show consistent initial loading spinner
This change makes sure that the loading spinner on the initial page load gets displayed consistently during the bootstrap of the client. The spinner gets hidden eventually after all apps have been announced as ready.

Also aligns the spinner design with all the other loading spinners and reduces the delay before showing it from 1 to 0.5 seconds.