refactor: use AppWrapper for preview app
Makes use of the `AppWrapper` component in the preview app. This reduces boilerplate code in the app and enables the context menu in the app bar.

The resource loading still happens in the preview app because it's heavily tied to it. The loading in the `AppWrapper` is being bypassed during this.