refactor: id-based webdav requests for listings
Introduce id-based webdav requests for space and folder listings. This is the recommended way to work with the webdav API according to oCIS because it delivers better performance, especially with slower storages.

Some of the webdav operations still need path handling though. Because of that, this change still "injects" full resources paths in the space loader. This can (hopefully) be removed once all operations work id-based or when we make the switch to Graph.