feat: target upload folder
Adds the possibility to specify a target folder for an upload.

This needs to be done by calling `setUploadFolder` in the `UppyService`. An `uploadId` needs to be provided in addition to the target folder. This id needs to be present in the `meta` object of the files that need to be uploaded.