fix: focus on sortable table headers
Sortable table headers can now be focussed properly, including an outline reflecting that.

Instead of focussing the sort icon, it now focusses a button that wraps both, the name and the icon. The problem with just focussing the icon is that it does not always show - only if the table is currently sorted by this column.