feat: add permissions to share spaces
Since share spaces are virtual, they and their permissions can't be retrieved from the server. Instead, we have to load the permissions via the underlying share. This change does exactly that and persists the permissions for the current user within the space object so they can be accessed and checked when needed.

Also makes sure that share spaces are being added to the spaces store so they're not just temporary objects.