fix: remove duplicated space requests in space roots
The duplicated requests occured because after listing files, we check the given `fileId` against the id of the root element. If they differ, we assume an incorrect path and therefore re-do the requests with the correct `fileId`.

This doesn't work for spaces though because their ids always differ from the given `fileId` (they are missing the node id as last part of the id). So the solution is to add an exception for spaces.