refactor: use request interceptors in client service
Removes the old way of re-initializing clients in the client service when they needed to. The client service now uses request interceptors for the dynamic headers such as `Authorization`.

This a) is more developer friendly since the client service can now be destructured and b) removes the overhead of re-initializing the clients over and over again.