chore(web-pkg): prepare publish with implementation
Prepares `web-pkg` for a publish with the actual implementation, not just the types. This is needed for unit tests in external apps, since the runtime is not present in this context to provide the actual implementation.

Also moves most of the peer dependencies to regular dependencies (expect for the monorepo packages) because it's unrealistic for an app that uses this package to provide all the peers.