chore(web-pkg): prepare publish with implementation
Prepares `web-pkg` for a publish with the actual implementation, not just the types. This is needed for unit tests in external apps, since the runtime is not present in this context to provide the actual implementation.

Also moves the peer dependencies to regular dependencies because the app that uses this package should not need to provide all the peers.