fix: folder size mismatch
Fixes the size mismatch between a folder in the file list and the accumulated size inside the folder beneath the file list.

The latter now takes the size of the current folder and only falls back to manually accumulating the size if no current folder given. Also removes the `totalResourcesSize` computed from the resource store because it's only used in one place and we don't want to encourage others to use this method of calculating the current folder's size.