fix: preserve language in password protected folders
This adds a new  URL query parameter which is used as the default language of the UI. It is ignored when the value does not match supported language codes or when user is logged in. This param is now used in folder view in password protected folders app to preserve the language when opening a folder.