fix(web-pkg): download archives directly
When in authenticated context or inside public links without password, download the archive file directly by opening the download URL instead of fetching it. This gets around constrains on memory. In public links with password, we still need to request the file first due to missing public link signed URLs. At least we are now requesting the file as blob instead of arraybuffer which works at least with bigger limits in Chrome and Firefox.