fix(web-pkg): address user feedback in markdown editor
This fixes several bugs introduced in the new markdown editor. Several dependencies were added to make features like mermaid, highlighting, fullscreen, and katex work. Links for keyboard shortcuts and supported languages in code blocks are added. Theme for mermaid is adjusted so that everything is visible in dark mode. Hide GH link to the package repo, save button in toolbar, and image upload.