chore: [OCISDEV-536] enable encoded chars for collabora
In the docker compose file, we now enable encoded characters in Traefik
for Collabora service. This is needed to fix proxy issues with Collabora
after updated the Traefik image version to 3.5.4.
We also changed the `o:net.frame_ancestors` to
`--o:net.content_security_policy` because the former configuration
option is deprecated.